Reenable move semantics and fix compiler warnings (#520)
The ``std::random_device`` instance is not stored as a member
variable on the ``RedisServer`` class. The ``std::random_device`` is not
movable or copyable, and as a result, it was implicitly disabling move
semantics for the ``Client`` and the ``RedisServer`` derivates.  Also,
``Exception`` class methods have been marked with the appropriate
``const override`` keywords to fix compiler warnings.

[ committed by @mellis13 ]
[ reviewed by @ashao @al-rigazzi ]